feat(redis): move examples and test into a separate Go module
This avoids having github.com/redis/go-redis/v9 as a dependency.
