+++ modules/redis/example/examples_test.go
@@ -1,4 +1,4 @@
-package redis_test
+package example
 
 import (
 	"context"
