+++ modules/redis/example/go.mod
@@ -1,16 +1,18 @@
-module github.com/testcontainers/testcontainers-go/modules/redis
+module github.com/testcontainers/testcontainers-go/modules/redis/example
 
 go 1.25.0
 
-toolchain go1.25.9
-
 require (
-	github.com/mdelapenya/tlscert v0.2.0
+	github.com/google/uuid v1.6.0
+	github.com/redis/go-redis/v9 v9.7.3
 	github.com/stretchr/testify v1.11.1
 	github.com/testcontainers/testcontainers-go v0.42.0
+	github.com/testcontainers/testcontainers-go/modules/redis v0.42.0
 )
 
-replace github.com/testcontainers/testcontainers-go => ../..
+replace github.com/testcontainers/testcontainers-go => ../../..
+
+replace github.com/testcontainers/testcontainers-go/modules/redis => ..
 
 require (
 	dario.cat/mergo v1.0.2 // indirect
@@ -24,6 +26,7 @@ require (
 	github.com/containerd/platforms v0.2.1 // indirect
 	github.com/cpuguy83/dockercfg v0.3.2 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
 	github.com/distribution/reference v0.6.0 // indirect
 	github.com/docker/go-connections v0.6.0 // indirect
 	github.com/docker/go-units v0.5.0 // indirect
@@ -32,10 +35,10 @@ require (
 	github.com/go-logr/logr v1.4.3 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/go-ole/go-ole v1.2.6 // indirect
-	github.com/google/uuid v1.6.0 // indirect
 	github.com/klauspost/compress v1.18.5 // indirect
 	github.com/lufia/plan9stats v0.0.0-20211012122336-39d0f177ccd0 // indirect
 	github.com/magiconair/properties v1.8.10 // indirect
+	github.com/mdelapenya/tlscert v0.2.0 // indirect
 	github.com/moby/docker-image-spec v1.3.1 // indirect
 	github.com/moby/go-archive v0.2.0 // indirect
 	github.com/moby/moby/api v1.54.2 // indirect
