+++ modules/redis/example/testdata/Dockerfile
