+++ modules/redis/example/testdata/redis6.conf
